Error post for front-end. Toggling login and new post
