+++ Markobacon_Frontend/src/MarksideComponents/MainColError.js
@@ -1,0 +1,9 @@
+function MainColError(props) {
+    return (
+        <>
+        <p className="homeTopBar">Something went wrong rendering this... And this needs styling.</p>
+        </>
+    )
+}
+
+export default MainColError;
