+++ Markobacon_Frontend/src/MarksideComponents/MainColumnContent.js
@@ -1,4 +1,5 @@
 import BlogFeed from "./BlogFeed";
+import MainColError from "./MainColError";
 import NewPost from "./NewPost";
 
 // TODO: Couldn't I just make this into a function? Maybe a component just feels nicer.
@@ -8,7 +9,10 @@ function MainColumnContent(props){
     if(props.content === 'latestPosts'){
         return <BlogFeed />
     } else if (props.content === 'newPost'){
-        return <NewPost />
+        // TODO: Do a double-check to ensure someone has the right token
+        const tempLoggedIn = false; // Temporary boolean for front-end testing
+
+        return tempLoggedIn ? <NewPost /> : <MainColError />
     }
 }
